Time: 1 ms (98.80%), Space: 39.6 MB (84.59%) - LeetHub

+++ 69-sqrtx/69-sqrtx.java
@@ -1,7 +1,7 @@
 import java.lang.Math;
 class Solution {
     public int mySqrt(int x) {
-        double a=(int)Math.sqrt(x);
+        double a=Math.sqrt(x);
         return (int)a;
     }
 }
